[jax2tf] Clean up the jax2tf sharding_tests.
The sharding tests are very important because jax2tf must do non-trivial manipulation of shardings, e.g., to wrap the inputs and the outputs with
sharding annotations.

Merged two test classes ShardedJitHloTest and Sharding. One of the classes was
just checking the annotations in the TF HLO, the other one was just running
the code and comparing the results. Now we do both in one test.

Refactored the code to log JAX HLO and to check the occurrences of annotations
in TF HLO. Now we support checking that the occurrence count is equal to a value
or greater or equal to a value.

Added more annotation checking. This makes the test more informative (because
there is no other good way to check that sharding was applied correctly).
But this makes it also possible that the test will fail when we change the
JAX lowering.

PiperOrigin-RevId: 518362978