[3/n] Introduce Module into FunctionDefLibrary.
Add an option `native_serialization_as_module` to jax2tf to serialize the stablehlo explicitly to `FunctionDefLibrary`.

In jax2tf native serialization, the stablehlo is embedded as a string attribute of the `XlaCallModuleOp`.

When `native_serialization_as_module` is `True`, the stablehlo is serialized as a `ModuleDef` in `FunctionDefLibrary`, then `XlaCallModuleOp` refers to the `ModuleDef` using the module's name as a string attribute.

PiperOrigin-RevId: 518850240