Use modern packaging standards for setuptools
In setuptools 30.3.0 (December 2016), support was added for specifying
setup options in setup.cfg as opposed to arguments to the setup
function. Options are merged, and setup.py takes priority when
disagreements occur.

setup.cfg arguments are superior for a couple reasons. They are:
- declarative -- no python code is executed to evaluate them
- easier to read
- capable of doing things that you'd need to open-code in python

Of particular note, it is possible to load files or statically parse the
AST for a simple assignment, via values such as `attr:` or `file:`. This
is a cfg-exclusive feature.