[jax2tf] Refactor jax2tf to move more pure-JAX stuff to jax_export
This is the next step in the refactoring and simplifying jax2tf. An additional goal is to move to jax_export.py as much pure-JAX functionalit as possible so that jax_export can start to become a standalone API for native serialization.

Previous PRs in this series:

  * [shape_poly] Refactor the computation of the dimension variables in native serialization #15258 and [shape_poly] Improved computation of dimension variables for native serialization #15287: refactoring computation of dimension variables
  * [jax2tf] Cleanup of handling of tf.custom_gradient #15341 and [jax2tf] Refactor the gradient machinery for native serialization #15432: refactor handling of tf.custom_gradient (moved VJP serialization to jax_export)

Here we do the following:

  * we move the handling of in/out pytrees to jax_export. This allows the removal of a workaround that introduced a trampoline for the lower function for native lowering.
  * we move most of the code to set up the input abstract values into jax_export (poly_spec and poly_specs)
we separate out the implementation details for graph and native serialization into two classes GraphSerializationImpl and NativeSerializationImpl. This simplifies the core of the jax2tf.convert.

This is a pure refactoring, it should not change any existing behavior for jax2tf.201~