[jax2tf] Fix backwards compatibility test to not use goldens for eigh
We now check the expected property for the eigenvectors, instead
of comparing them to the goldens. We still compare the eigenvalues
to goldens. The check is a copy of the similar check in linag_test.py:
https://github.com/google/jax/blob/6b7ae36f1087866fdd92c29ce2f232f2410943e7/tests/linalg_test.py#L1653

Also discovered a fixed that we were not passing the `rtol` parameter
correctly to tests.

PiperOrigin-RevId: 523762205