Jax101 - Clarify that the compiled code is executed on first call
The current wording implies that the first time a jitted function
is called, the computation happens in Python. It's actually only
the tracing that happens in Python, and the compiled code is run
during the first call. The distinction is important e.g., to
understand why it might make sense to jit a function that's only
called once.

PiperOrigin-RevId: 526906176