[jax2tf] Fix float0 representation under native serialization
Previously, jax2tf used int32 zeros for float0 while the JAX
native lowering uses bool. The actual type does not make any
difference for graph serialization because TF will not use
the values. But for native serialization this meant that we
passed wrong-dtype arguments to the exported modules.

This error was masked by the shape refinement logic in XlaCallModule,
which set the type of the module formal arguments to be the
same as the type of the actual arguments.