[shape_poly] Refactor the unification of the argument abstract values with the actual arguments
This was called shape_poly.compute_dim_values. We rename it to
shape_poly.unify_avals_with_args and we add better error reporting to it.
Now it will identify the arg/kwarg where there is a shape discrepancy.

This is intended to be a pure refactoring, in preparation for adding
support for shape polymorphism to jax_export.call_exported.