Modify the QDWH algorithm to run subspace iteration on the projector of smallest rank.
The QDWH splitting step involves two orthogonal projectors
P_plus = -0.5*(U-I) and P_minus = 0.5*(U+I), one of which will have rank k and the other rank n-k. Ideally, if we are able to pick the median eigenvalue for the split point optimally, k will be near n/2, and the rank of the two projectors will be similar. However, if our guess of the median eigenvalue is poor or the matrix is rank-deficient, k can be far from n/2, and the cost of the subspace iteration will be higher for the projector of higher rank, since it involves computing the QR decomposition of a matrix of size n x rank.

This change makes the algorithm adaptively pick the projector of lower rank.

PiperOrigin-RevId: 528941151