Improve the bisection strategy in eigh.
Background: Currently, we pad the sub-matrices that occur during the spectral bisection algorithm to fit in a small number of buckets, in order to keep compilation time down. Each unique bucket size gives rise to a separate JIT compilation. The current strategy uses powers of two times the termination size of 256, below which we switch to a Jacobi solver. One issue is that the bisection step rarely splits the matrix in two exact equal parts, so one of the child-problems is forced to use the large bucket size of its parent, which wastes significant device cycles.

This changes modifies the bucket selection strategy to not use [256, 512, 1024, ... n], but instead include a little slack at each level, such that both sub-problems from a non-perfect split will likely fall into the smaller bucket size. Specifically, we add 4% slack and round up to the next larger multiple of 32. These heuristic values were found experimentally. As an example, for n = 2048, we get the bucket sizes [2048, 1056, 544, 288, 256].

Maasuring runtimes on random matrices of size 512, 1024, and 2048, we see significant speedups:

   N |  wall time before | wall time after
===========================================
 512 |       27.8 ms     |      24.8 ms
1024 |       97.6 ma     |      79.3 ms
2048 |      414.5 ms     |     308.0 ms

PiperOrigin-RevId: 529567005