add a jax.nn.softmax_alt with alternative differentiation behavior
Following up on #15677, we found that at least one user's code worked better
with the old differentiation approach. So we decided to keep it alive! Instead
of relying on a flag, users can just apply `jax.nn.softmax_alt` to get the old
behavior. (The flag is still here for now, though I'll delete it in follow-up.)