Fix warnings & incorrect return value in dtypes.assubdtype
This fixes an error introduced in #16010. The problem is that if b is a np.dtype object, then `b in [np.floating]` uses np.dtype.__eq__, which converts its argument to a dtype and leads to a warning. It also returns incorrect results in the case `b = np.float64`, because `np.dtype(np.floating)` returns float64, and thus `b in [np.floating]` evaluates to True.

This is why the original code treated the `issubdtype(b, np.dtype)` condition separately from the list membership check, but the problematic case was previously untested.

PiperOrigin-RevId: 532554190