[PJRT PLUGIN] Add automatic discovery of PJRT plugins from pip installed packages.
The plugins in the namespace package `jax_plugins` will be imported. The plugins need to (1) be placed in a root folder `jax_plugins` and follow other namespace package requirements, and (2) implement an initialize() method which appends `plugin_name:file_path` to env var `PJRT_NAMES_AND_LIBRARY_PATHS`.

Appending to PJRT_NAMES_AND_LIBRARY_PATHS is a short term solution and what the initialize() should do is in discussion.

PiperOrigin-RevId: 532897890