Some cleanup of logging in XlaCallModule.
Previously, we were using XLA_VLOG_LINES to ensure that the
whole module is logged even when large. This had the unpleasant
side-effect that each line of the module was logged separately,
including the filename and line number. This made the log large
and prevented copy-pasting the module.

Now we use `if (VLOG_IS_ON()) LogModule`.

PiperOrigin-RevId: 536958739