[jax2tf] Add a backward compatibility test for tf.call_tf_function
`tf.call_tf_function` arises from `jax2tf.call_tf(tf_fun, call_tf_graph)`. However, a function that contains this can be lowered and executed only with `jax2tf.convert` and ought to be serialized as ` tf.Graph` because the serialization includes a tf.function as well.

In order to support this we need to add some code to back_compat_test.py to serialize and run the serialized code as tf.Graph.

PiperOrigin-RevId: 537062963