[shape_poly] Enable calling from JAX modules that use shape polymorphism.
If a JAX function uses shape polymorphism the resulting StableHLO contains
dynamic shapes and it is not directly compileable. However, in such modules
the dynamic shapes depend only on the input shapes, and in JAX jit the input
shapes are static. So, we run a shape refinement pass over the module to
resolve the dynamic shapes prior to compilation.

PiperOrigin-RevId: 538275268