Revive indirection of RaggedAxis batch dims.
There are boundaries across which segment lengths should flow as
explicit, object-level arguments, instead of as "metadata".  To cross
such a boundary, we need to be able to extract the embedded segment
lengths from a RaggedAxis object, replacing them with references to
known object-level arguments; and we need to be able to reconstruct
the RaggedAxis object as metadata by resolving those references.