Do not register the "interpreter" backend by default.
This is a dummy backend useful only for debugging that uses XLA's HLO interpreter. It's doubtful many people care about this or even know it exists.

During initialization this plugin spawns two thread pools, which may allocate 2x# cores threads, which is wasteful if they are never used. This may contribute to some users hitting thread limits on their machines.

PiperOrigin-RevId: 539107641