Checkify: release memory by deleting references to variables after last use.
This is a memory optimization that can be applied to all jaxpr interpreters that
execute op-by-op computation (ie. computation that has not been staged out to
compilation). We loop through the equations of the jaxpr and write/read
variables from an environment. The environment currently holds a reference to
variables throughout the whole evaluation of the jaxpr, even after the variable
is no longer needed. When the variables are backed by large on-device buffers
(which is the case when executing op-by-op) this results in much higher memory
peaks than the user would expect.

This change adds an optimization which removes all variables that have been used
for the last time from the environment after every equation execution. This
optimization should be factored out and added to each jaxpr interpreter, which
will be done in a follow-up.

Note that the most-used jaxpr interpreter, which is reverse-AD, does not have
this issue because it pops every cotangent variable it reads!
(https://github.com/google/jax/blob/9816a6a905d798faad44b3f7aa1dfe8f5ca66433/jax/_src/interpreters/ad.py#L208)

PiperOrigin-RevId: 539942844