[shape_poly] Fix the lowering of lax.dynamic_shapes with shape poly
lax.dynamic_slice clamps the start indices to ensure the access
is in bounds, but when lowering with shape polymorphism we use
stablehlo.RealDynamicSliceOp, which is a version of SliceOp
and does not have the clamping. We clamp explicitly during lowering.

This changes the lowering in very limited circumstances: when we have
a lax.dynamic_slice with shape polymorphism and a dynamic slice size,
under native lowering only, and only when the start indices were out
of bounds.