Don't allow the use of stale cache entries in Github actions CI.
The use of restore-keys allows the pip cache to match stale cache entries. This allows for partial cache reuse, but means the cache also grows without bound.
At the time of writing, the pip cache is 4.7GB, which is much larger than just downloading the wheels we need (20MB)! So it's probably a net positive not to reuse caches even if it requires some redundant downloads.