Removed the type annotation of `orig_out_sharding_handlers`
The annotation was a generic type alias which are not well supported for
pytype and can, in some cases, lead to crashes.

It was used "raw" and was thus implicitly specialized to `Any`. So, the
constraint of `SubClassT` being the same in the key, parameter and return
types was not actually enforced. It is now enforced through the `_register_out_sharding_handler` function.

PiperOrigin-RevId: 545460402