accept general `shape` option in `jax.random.split`
Several PRNG implementations (notably partitionable threefry) support
splitting to arbitrary shapes, rather than only to a 1-D vector of
keys. This change:

* Upgrades `jax.random.split` to accept a general shape as an
  argument.
* Updates the internal PRNG interface, and our various PRNG
  implementations, to accept and handle such a shape argument.

This change keeps the argument name `num`. We can still think on
whether and how we'd like to upgrade to `shape`.

Note that we could have supported arbitrary shapes by reduction to the
previous API (with a flat split count), using reshapes. We'd like to
avoid that, so as not to hide this structure from the underlying
implementation. For instance, partitionable threefry hashes a *shaped*
iota in order to split keys, and we don't want to flatten and reshape
around that for no reason.

Co-authored-by: Jake Vanderplas <jakevdp@google.com>