[shape_poly] Add support for max0 for symbolic dimensions.
There are a few cases when JAX computes `max(v, 0)`, most
notably when computing the sizes of strided access,
dilated convolutions and padding, and for the size
of jnp.arange.

Until now these cases were supported
for shape polymorphism only when we can tell statically
that the size is >= 0. Here we add support to the
symbolic expressions for a `non_negative` operator,
which essentially implements `max(v, 0)` and with this
we can now support the general case for `jnp.arange`, with
simpler code.

We could add a general `max` operator, and we may do so in the
future, but for now `non_negative` suffices.

Note that this fixes a couple of bugs

  * for core.dilated_dim we had the code "if d == 0 then 0 else ..."
  but this works only if we can tell statically that `d == 0`, and
  it produced wrong results when `d` was symbolic and could take
  the value 0.
  * for core.stride_dim we did not handle correctly the case when
  `d < window_size`.

Handling the above fundamentally requires a `max(d, 0)` operation.