[shape_poly] Improve error messages for shape assertions
Starting with serialization version 7 we introduce shape
assertions that are checked at runtime. In the process of
rolling out version 7 we encoutered projects with failed
shape assertions and it became clear that we need better
error messages.

See the changes here in tests and README.md for example of
the updated assertions.

To produce these assertions we now pass multiple operands to
the shape assertion, and we introduce a CachedShapeEvaluator
to reduce the amount of duplicate code generated.